+++ tests/test_pytorch/test_deepspeed.py
@@ -40,6 +40,7 @@ from lightning_habana.pytorch.accelerator import HPUAccelerator
 from lightning_habana.pytorch.plugins import HPUDeepSpeedPrecisionPlugin
 from lightning_habana.pytorch.strategies import HPUDeepSpeedStrategy
 from lightning_habana.pytorch.strategies.deepspeed import _HPU_DEEPSPEED_AVAILABLE
+from lightning_habana.utils.resources import get_device_name_from_hlsmi
 
 if _HPU_DEEPSPEED_AVAILABLE:
     from deepspeed.runtime.activation_checkpointing.checkpointing import checkpoint
@@ -199,7 +200,7 @@ def test_hpu_deepspeed_strategy_env(tmpdir, monkeypatch, deepspeed_config):
         pytest.param(
             "fp8",
             marks=pytest.mark.skipif(
-                HPUAccelerator.get_device_name() == "GAUDI", reason="fp8 supported on Gaudi2 and above."
+                get_device_name_from_hlsmi() == "GAUDI", reason="fp8 supported on Gaudi2 and above."
             ),
         ),
     ],
@@ -458,7 +459,7 @@ def test_multi_optimizer_with_hpu_deepspeed(tmpdir):
         0,
         1,
         2,
-        pytest.param(3, marks=pytest.mark.skipif(HPUAccelerator.get_device_name() == "GAUDI2", reason="Not supported")),
+        pytest.param(3, marks=pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI2", reason="Not supported")),
         "infinity",
     ],
 )
@@ -735,9 +736,7 @@ def test_lightning_deepspeed_inference_kwargs(enable_cuda_graph, device_count):
         torch.float,
         pytest.param(
             torch.float16,
-            marks=pytest.mark.skipif(
-                HPUAccelerator.get_device_name() == "GAUDI", reason="FP16 is not supported by Gaudi1"
-            ),
+            marks=pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI", reason="FP16 is not supported by Gaudi1"),
         ),
     ],
 )
@@ -770,9 +769,7 @@ def test_lightning_deepspeed_inference_params(device_count, dtype):
         torch.float,
         pytest.param(
             torch.float16,
-            marks=pytest.mark.skipif(
-                HPUAccelerator.get_device_name() == "GAUDI", reason="FP16 is not supported by Gaudi1"
-            ),
+            marks=pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI", reason="FP16 is not supported by Gaudi1"),
         ),
     ],
 )
@@ -805,7 +802,7 @@ def test_lightning_deepspeed_inference_config(device_count, dtype):
 
 
 @pytest.mark.parametrize("stage", [1, 2, 3])
-@pytest.mark.skipif(HPUAccelerator.get_device_name() == "GAUDI", reason="fp8 / fp16 supported on Gaudi2 and above.")
+@pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI", reason="fp8 / fp16 supported on Gaudi2 and above.")
 def test_hpu_deepspeed_training_accuracy(tmpdir, device_count, stage):
     """Test compare training accuracy between bf16 and fp8 precision for deepspeed."""
 
@@ -870,7 +867,7 @@ def test_hpu_deepspeed_training_accuracy(tmpdir, device_count, stage):
 
 
 @pytest.mark.standalone_only()
-@pytest.mark.skipif(HPUAccelerator.get_device_name() == "GAUDI", reason="Accessory test for fp8 inference.")
+@pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI", reason="Accessory test for fp8 inference.")
 def test_hpu_deepspeed_bf16_inference_accuracy(tmpdir, device_count):
     """Test maintain bf16 test loss used in fp8 inference accuracy test using deepspeed."""
 
@@ -915,7 +912,7 @@ def test_hpu_deepspeed_bf16_inference_accuracy(tmpdir, device_count):
 
 @pytest.mark.standalone_only()  # HQT cannot be reconfigured in same process
 @pytest.mark.parametrize("quant", [False, True])
-@pytest.mark.skipif(HPUAccelerator.get_device_name() == "GAUDI", reason="fp8 supported on Gaudi2 and above.")
+@pytest.mark.skipif(get_device_name_from_hlsmi() == "GAUDI", reason="fp8 supported on Gaudi2 and above.")
 def test_hpu_deepspeed_fp8_inference_accuracy(tmpdir, device_count, quant):
     """Test compare inference accuracy between bf16 and fp8 precision for deepspeed."""